requirements.txt
kucoin-python==1.3.9
requests==2.31.0
